xiufu

+++ src/main/java/com/oddrock/common/file/FileUtils.java
@@ -510,7 +510,7 @@ public class FileUtils {
 	public static File copyDirToParentDir(File srcDir, File dstParentDir) throws IOException {
 		File dstDir = new File(dstParentDir, srcDir.getName());
 		dstDir.mkdirs();
-		if(srcDir.listFiles()!=null) {
+		if(srcDir.exists() && srcDir.isDirectory() && srcDir.listFiles()!=null) {
 			for(File file : srcDir.listFiles()) {
 				File dstFile = new File(dstDir, file.getName());
 				copyFile(file.getCanonicalPath(), dstFile.getCanonicalPath());
